Significantly optimise CSS selection performance.
Now we pass a node bloom filter to css_get_style.  That node bloom
filter is filled with the node's ancestor element, class, and id
names.  Internally, libcss also generates a bloom filter for each
selector chain.  If the selector chain's bloom filter is not a
subset of the node bloom filter, we know that the selector chain's
rule does not apply to the node.  This avoids the slow selector
chain matching process.

Other smaller optimisations to move the ruling out of selector
chains for inapplicable media types and other reasons to before
we start comparing rules from different sources to find the next
rule.  All this is now done in hash.c so select.c never sees the
trivially ruled out rules.