Change how presentational hints are handled.
Previously, we performed normal selection from CSS sources, and then iterated
over all the properties in the populated computed style.  If the properties were
unset or their values were not from either a UA stylesheet or user stylesheet
with !important set, then we asked the client program (e.g. NetSurf) if there
were any presentational hints for that node, for each such property.

In the worst case this triggered N_PROPERTIES * N_NODES calls back to the client
program, even for properties that can't be set via HTML attributes.

The new API asks the client to supply a list of all the presentational hints
that apply to the given node.  For most nodes on modern documents, this is 0.
Any presentational hints are applied before selection from CSS sources.