HACKS for testing Media Queries parsing.
Top level stylesheets need to have their media query passed in.
So we need a way to parse standalone media queries.

This was hacked together to explore doing that.

However, it encounteded an issue where it seems the parseAtRule()
function in src/parse/parse.c doesn't handle the full grammar for
media queries.