Add end-block-content parse event
In order to cope with a situation where a block ends with a
selector which has no ruleset, add an end-block-content event
and in handling it, pop any intermediate states off the language
stack so that we're in block mode by the time the event is completed.

This fixes an assert situation caused by a ruleset such as:

  @media screen { dodgy } .outer { top: 10px; }

Which has been encountered in the wild (likely a typo).

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>